Crate pallet_randomness

Randomness Pallet

This pallet provides access to 2 sources of randomness:

local VRF, produced by collators per block

relay chain BABE one epoch ago randomness, produced by the relay chain per relay chain epoch These options are represented as type::RequestType.

There are no extrinsics for this pallet. Instead, public functions on Pallet<T: Config> expose user actions for the precompile i.e. request_randomness.

Local VRF

The local VRF randomness is produced every block by the collator that authors the block.

This pallet is default configured to look for the VrfOutput in frame_system::digests(). If it cannot find the VrfOutput in frame_system::digests(), then it will panic and the block will be invalid.

Next, the VrfOutput is verified using the block author’s VrfId and the VRF input, which is last block’s VrfOutput. If verification fails, then it will panic (block is invalid).

Finally, the output is transformed into the randomness bytes stored on-chain and put in LocalVrfOutput. Any pending randomness results for this block are filled with the output randomness bytes.

The function which contains this logic is vrf::verify_and_set_output. It is called in every block’s on_initialize.

Babe Epoch Randomness

Babe epoch randomness is retrieved once every relay chain epoch.

The set_babe_randomness_results mandatory inherent reads the Babe epoch randomness from the relay chain state proof and fills any pending RandomnessResults for this epoch randomness.

Config::BabeDataGetter is responsible for reading the epoch index and epoch randomness from the relay chain state proof. The moonbeam GetBabeData implementation is in the runtime.

Re-exports

pub use pallet::*;

pub use types::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

types

vrf

VRF logic

weights

Autogenerated weights for pallet_randomness

Traits

GetBabeData

Read babe randomness info from the relay chain state proof