Trait pallet_randomness::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type AddressMapping: AddressMapping<Self::AccountId>;
    type Currency: Currency<Self::AccountId>;
    type BabeDataGetter: GetBabeData<u64, Option<Self::Hash>>;
    type VrfKeyLookup: KeysLookup<NimbusId, VrfId>;
    type Deposit: Get<BalanceOf<Self>>;
    type MaxRandomWords: Get<u8>;
    type MinBlockDelay: Get<Self::BlockNumber>;
    type MaxBlockDelay: Get<Self::BlockNumber>;
    type BlockExpirationDelay: Get<Self::BlockNumber>;
    type EpochExpirationDelay: Get<u64>;
}

Configuration trait of this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Overarching event type

type AddressMapping: AddressMapping<Self::AccountId>

Address mapping to convert from H160 to AccountId

type Currency: Currency<Self::AccountId>

Currency in which the security deposit will be taken.

type BabeDataGetter: GetBabeData<u64, Option<Self::Hash>>

Get the BABE data from the runtime

type VrfKeyLookup: KeysLookup<NimbusId, VrfId>

Takes NimbusId to return VrfId

type Deposit: Get<BalanceOf<Self>>

The amount that should be taken as a security deposit when requesting randomness.

type MaxRandomWords: Get<u8>

Maximum number of random words that can be requested per request

type MinBlockDelay: Get<Self::BlockNumber>

Local per-block VRF requests must be at least this many blocks after the block in which they were requested

type MaxBlockDelay: Get<Self::BlockNumber>

Local per-block VRF requests must be at most this many blocks after the block in which they were requested

type BlockExpirationDelay: Get<Self::BlockNumber>

Local requests expire and can be purged from storage after this many blocks/epochs

type EpochExpirationDelay: Get<u64>

Babe requests expire and can be purged from storage after this many blocks/epochs